Add (text) map16 and level(s) extractables
The source addresses in exported MWL files are stripped by default. Any addresses whose bank bytes is 0xFF are left alone, anything else is set to 000000. Contrary to the current documentation on the MWL format, it seems like setting this pointer to 000000 for palette data does not actually cause Lunar Magic to think that the MWL doesn't use a custom palette, not sure what's up with that, but good for us, I guess.