Export Graphics and ExGraphics if in build order but no folders exist
Idea is that if you have no (Ex)Graphics folder(s), you probably want them, if they're in your build order. But if you *do* have them, you probably *don't* want to extract (Ex)Grahpics, cause the files in the filesystem are likely at least as up to date as the ones in the ROM, which makes it dangerous to overwrite them. Seems reasonable to me, but who knows.