Allow saving a clean ROM if no built ROM is available
Mostly so you can get started easily if you're starting a fresh project. Design is not ideal since you can accidentally overwrite previously exported data, but it's hidden behind a safety prompt and pretty clearly indicated, so honestly, if you have this happen to you that's on you, is my take on this one.