Fix a heap corruption crash (??)
Not even sure whether this fixes anything, the bug kinda seems to come and go, only in release mode and only sometimes. Very fun.

It just kinda crashed during or after the build process, no clue what was/is actually causing it. I thought this might have fixed it (maybe taking a reference of a smart pointer was not a good thing to do?), but it also seems to no longer happen if I undo this commit and compile the previous one, so who fucking knows.