Ensure we can tell if ROM's last write time has changed
Used to be the case that if the ROM's last write time changed, but the marker string remained intact, we wouldn't be able to tell that the ROM might have changed. 

I'm not storing the ROM's last write time inside the marker string, so that I can later tell if they still match or if the last write time has changed, indicating that we should export resources before building.