Rename project to "callisto"
There was already a hack named "Stardust" and also it's not a particularly fun name. 

Callisto is both a gay nymph in Greek mythology, who gets turned into a bear after fucking Zeus who is disguised as Artemis, as well as an actual moon, so I think that's a neat fit.