Ensure ROM's last modified time matches cached timestamp
Turns out you could still trick callisto into not noticing ROM changes by keeping a "backup" ROM with a marker inside it that matches its last modified time and replacing the current ROM with this older ROM.

To work around this, I now store the "last sync time" in the callisto cache and check if it matches both the ROM's last modified time as well as the internal marker's timestamp to determine if it has changed since the last sync.

Should now detect basically any changes that you can make to the ROM in any way without really going out of your way to fuck with it (I hope).