Speed up level export drastically on multicore machines
This is notably a disgusting process where I create one copy of the ROM per CPU core and edit each of them to make only a certain smaller chunk of levels appear as "modified" to LM, so that when I then instruct it to export all "modified" levels, it only exports the chunk the corresponding core is responsible for.

It is probably a good idea to make this optional, as it could probably break in the future once fusoya decides it's time to change this behavior out of spite or on accident.

Notably will still need to update error handling a bit (for some of the previous commits as well), since throwing exceptions in threads is not generally ideal unless they're caught.