Change init thread execution a bit
Basically, during each insertion step, we also perform the init step for the next insertion in parallel. 

This doesn't really matter for most stuff, but for FLIPS insertables this lets me already prep the next ROM during the current one, which might help a bit (it's not very noticeable I think). 

Note that I unfortunately seem to not be able to insert (ex)graphics in parallel, seems like LM accessing the files might make them inaccessible for the duration or something, not sure exactly.