Multithread conflict tracking and track initial patch conflicts
Don't know how beneficial the multithreading actually is, but I guess I might as well keep it?

Accounting for the initial patch in conflict tracking seems to really inflate the number of detected conflicts, probably mostly false positives, but I guess it's easy enough to filter out stuff you don't care about so might as well leave this in for now.