Ensure level filenames conform to format prior to build
Turns out with Lunar Helper (and callisto, by extension), it was possible to place a MWL like `level 7.mwl` in the levels folder, then build and have that level inside the ROM, but have it be exported by Lunar Monitor/Helper as `level 007.mwl` later on, which would cause there to be two distinct levels both being inserted as level 7, one after the other, the latter overwriting the earlier one.

 I'm now working around this by ensuring all files in the levels folder follow the `level ###.mwl` format prior to builds. The way I do this is:

- if the filename doesn't conform to the format, I ask the user if they want it to be renamed to `level X.mwl` where `X` is the level number the MWL points to (I do have to lightly inspect the MWL's contents for this)

- if the filename *does* conform to the format, but the level number in the filename doesn't match the one in the MWL, I ask the user if they would like to either rename the file or point the file to a different level

The first is likely to happen when people drop levels straight into the levels folder from external sources. The latter occurs when people think they can just create a copy of the MWL and rename it, neglecting to consider that the filename actually has nothing to do with the level number from Lunar Magic's perspective.

Note that I currently take user input on these prompts with enter, which has the unfortunate effect of now throwing the user out of the command line view immediately when the build process finishes. I hate all this "press enter"/"press ESC" shit and wish to be done with it, please and thank you. Maybe I'll just scrap FTXUI and go back to a plaintext menu, this shit's lowkey real annoying.