Rework modules to use `input_path` and `output_paths` options
`input_path` should point to the source code/data for the globule, while `output_paths` is *optionally* a list of paths at which files containing the addresses of the inserted labels will be made available by callisto (these paths are relative to the main directory callisto keeps inserted modules in: `.callisto/modules`).

If `output_paths` is omitted, the default will be used, which is just the name of the source module + ".asm" extension in the root of the modules folder.

Note that the filename used in the `output_paths` will directly influence the labels made available in the output file, i.e. naming an output `my/cool/output.asm` will prefix the labels in the output file with `output`, even if the input file is differently named.