Automatically reload ROM at end of build and quickbuild
Note that right now, the ROM is reloaded no matter what (no work to do for quickbuild, errors during builds, etc. are ignored). But this should be easy enough to change.

The mechanism for this is highly funky and should be a ton more reliable than LH's shitty way of doing it, but I could still probably do some more error handling, just to be sure that things go smoothly.

Bringing LM to top should also be really simple from here, pretty much just need the code that brings the "main" window of a PID to top.

Closes #54