Implement automatic resource exports and no longer leak eloper process
Ok, so you kinda have to give it a few seconds to finish exporting and there's not really any indication that there's a save going on, but it *does* work. Spamming a bunch of saves really quickly might be an issue, but nobody does that anyway and it might also actually be fine, who knows.

One good thing is that the save still continues even if you close LM before the save has finished, so that's nice.

I do fear that if people have really huge hacks that take a long time to save, they could hit Q or R in callisto before the save from Eloper has finished, but that could potentially be worked around by passing a bool through shared memory indicating that a save is still going on. Or maybe use a timestamp and if it takes suspiciously long you can assume that the Eloper Save has somehow ran into an issue and just ignore it.

Overall it's a little scuffed still, but it does work.

Notably still missing is re-attaching to previously injected Lunar Magic instances, but that should probably not be too bad to implement.

Oh, profiles are also handled not that well right now. Eloper just picks the first profile for use if there are any, which might be an issue for weirdos who use different save locations or something depending on their profile, but I'm not gonna handle that right now because hopefully nobody is doing that (read: someone will be doing that for some reason).