Re-attach to injected Lunar Magic instances on startup
Right now, callisto will look for processes using the Lunar Magic executable at `lunar_magic_path` and for each of those check if a shared memory region exists. If so, it will re-attach itself to this instance.

This should allow callisto to still account for these processes automatically exporting resources as well as to still reload the ROM in those instances.

Note that moving around the Lunar Magic executable, changing `lunar_magic_path` or starting callisto with an invalid configuration may cause the re-attachment to not work correctly. Not sure what I will do about that, if anything.

Closes #59