Get rid of `temporary_folder`
We now just make a temporary folder at whatever the OS's temporary folder + "callisto_build_temp" is. The folder is freshly deleted and remade on build start and deleted on build success, but not on build failure, so that users can still inspect build artifacts that ran into issues if they want.

Closes #73