Ensure different prompts in multithreaded save don't prompt at same time
This *may* be the cause of a hard crash in combination with the YYCHR issue.

The idea is that the YYCHR issue might cause saves to abort halfway, then leaving (Ex)GFX folders in an intermediate state, where they differ from what's in the ROM.

Attempting a save after this (probably prompted due to "unexported resources") would then cause both prompts at roughly the same time, which is apparently UB if it's in a multithreaded environment. The added mutex around the prompt should be a bandaid on this.

I was never able to reproduce this crash on my own PC, but I'm hoping this fixes it for others.