Fix modified level calculation & levels-per-thread calculation in save
Big thanks to @blackmagemathie for helping me debug this!

I was checking if a level is modified by checking if the bank byte of the layer 1 pointer is >= 0x10, which is not good enough, because the high byte also needs to be >= 0x80. This could cause Callisto to overestimate how many modified levels exist in the ROM, making it sometimes choose to use multithreading when it was not actually likely to lead to faster export times.

Furthermore, my calculation of the number of levels each thread should export turned out to be not exactly sound math, which could result in the following:
- undefined behavior (due to reading past the end of a vector)
- an endless loop creating an ever-growing ROM (not great, the biggest spotted in the wild was 40 GB, probably not playable)

The reason most people probably never ran into this is that whether the bug occurs depends both on the number of available threads as well as the number of modified levels in your ROM (or rather, how many modified levels Callisto *thought* you had in your ROM). This is also why the bug was essentially impossible to reproduce for me.