Let patches and modules know they're patches or modules
This might sound silly, but this lets you do things such as insert the same file as either a patch or a module, issue a warning or error to the user if they inserted the resource as the wrong thing or do something else entirely depending on the insertion type, which is maybe kinda cool.

For patches, `!CALLISTO_INSERTION_TYPE` will be `Patch`, for modules it will be `Module`.