Make `call_module` no longer clobber A
Should now also work regardless of whether A is 16 or 8-bit.

Shoutout to karafruit for suggesting this particular technique, I think it's quite nice. Not sure if there's one that's faster and as convenient, but if you're aware of one open a PR.

Note that this does make the macro 7 cycles slower, from 21 cycles to 28, but definitely worth the convenience.

Closes #92