Pass path to `.callisto` directory to all tools as include path
A `.callisto` file containing the absolute path to the project's `.callisto` folder is now created before and deleted after each tool's run in its declared working directory. This is then read by the custom Asar implementation, which adds this folder as an additional include directory.

This now allows all tools and resources that use Asar to simply call `incsrc "callisto.asm"` instead of having to use a clunky relative path to the project's root. Should greatly aid portability of modules and ease development.