Actually receive level deletion message
Note that "deleting" levels from the ROM is gonna be weird behavior, cause all it really seems to do is make the level a test level and mark it as being unmodified. So Callisto won't export that level, which means it'll turn back into its vanilla version sooner or later...

Probably just tell people to either not use LM 3.50+ or to not bother deleting levels cause it doesn't work right.

Really not feeling like changing to exporting all levels just because of this, would be a waste of time and space.