Add `prefer_user_clean_rom` config variable
This config variable, assumed to be false by default, may only be specified in user settings and if set to true will ensure that if `clean_rom` is set in user settings it will not be overwritten by project or profile configurations. This is useful if you already have your clean ROM specified in your user settings and don't want to conform to project's expectations of where clean ROMs should be located, i.e. when using someone else's baserom.